Add our own method for determining the bounding box
I found some pdfs where pdfcrop/ghostscript determined
the bounding box incorrectly. With this commit we introduce
a simple way to determine the bounding box by finding the
smallest rectangle that is completely surrounded by white
pixels. This will work well for most PDFs. The centering
functionality is also improved, and now correctly centers
the PDF on the reMarkable both vertically and horizontally.