Fix provider selection for redirected urls
Some urls, such as the arXiv urls with the : in the identifier,
didn't work when using the UI interface because the redirected
url wasn't past to the provider, but the original url was. This
commit fixes that issue and adds unit tests for the provider
selection function, hopefully making this more robust in the
future.