Fix the pdf output of dearxiv
The dearxiv function removes the arXiv stamp from the pdf, but
in the past would return a somewhat broken pdf. When it
deleted the stamp the positions of objects and the length of the
stream of the arXiv stamp object would be incorrect. This wasn't
an issue for pdftk but gives a warning for qpdf. Because we don't
want to mask qpdf warnings in general (something might actually
be wrong), it was desirable to return a valid pdf file from dearxiv.
This commit does exactly that.