Replace excepthook for nicer errors
I think that generally a traceback is too much info for an
end-user, and the error message we provide can get lost in
the noise. This commit disables the standard excepthook
with its traceback, and replaces it with just the message
of the exception. Using the debug flag enables the original
traceback.